fix numpy deprecation

+++ co_calibration_gibbs_sampler_expressions.py
@@ -104,12 +104,12 @@ def posterior_sigma_y_explicit(sigma_y, a, b, Xa, Y, shape_sigma_y, scale_sigma_
             res = minimize(evaluate, x0=mode_sigma_y, bounds=((loc_sigma_y, None),), method="TNC")
         
         # use yet another optimization strategy (global) if previous result(s) unplausible
-        if np.float(res.x) <= loc_sigma_y:
+        if np.double(res.x) <= loc_sigma_y:
             evaluate = lambda x: np.square(target_quantile - quad(posterior_sigma_y_explicit_faster, loc_sigma_y, x, args=tuple(args))[0])
             #res = basinhopping(evaluate, x0=mode_sigma_y)
             res = dual_annealing(evaluate, bounds=((loc_sigma_y, loc_sigma_y+10),))
 
-        return np.float(res.x)
+        return np.double(res.x)
     else:  # return pdf at value sigma_y
         return posterior_sigma_y_explicit_faster(sigma_y, *args)
